// src/types/resources/InsurancePlan.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TMeta } from '../partials/Meta';
import { TUri } from '../simpleTypes/Uri';
import { TNarrative } from '../partials/Narrative';
import { TResourceContainer } from '../simpleTypes/ResourceContainer';
import { TExtension } from '../partials/Extension';
import { TIdentifier } from '../partials/Identifier';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TPeriod } from '../partials/Period';
import { TReference } from '../partials/Reference';
import { TInsurancePlanContact } from '../partials/InsurancePlanContact';
import { TInsurancePlanCoverage } from '../partials/InsurancePlanCoverage';
import { TInsurancePlanPlan } from '../partials/InsurancePlanPlan';

export type TInsurancePlan = {
  resourceType?: string;
  id?: string;
  meta?: TMeta;
  implicitRules?: TUri;
  language?: string;
  text?: TNarrative;
  contained?: TResourceContainer[];
  extension?: TExtension[];
  modifierExtension?: TExtension[];
  identifier?: TIdentifier[];
  status?: string;
  type?: TCodeableConcept[];
  name?: string;
  alias?: string[];
  period?: TPeriod;
  ownedBy?: TReference;
  administeredBy?: TReference;
  coverageArea?: TReference[];
  contact?: TInsurancePlanContact[];
  endpoint?: TReference[];
  network?: TReference[];
  coverage?: TInsurancePlanCoverage[];
  plan?: TInsurancePlanPlan[];
};